+++ src/components/results/reject/index.tsx
